+++ packages/codegen-ui-react/lib/react-theme-studio-template-renderer.ts
@@ -53,6 +53,7 @@ export class ReactThemeStudioTemplateRenderer extends StudioTemplateRenderer<
   ReactOutputManager,
   {
     componentText: string;
+    declaration: string | undefined;
     renderComponentToFilesystem: (outputPath: string) => Promise<void>;
   }
 > {
@@ -86,6 +87,7 @@ export class ReactThemeStudioTemplateRenderer extends StudioTemplateRenderer<
 
     return {
       componentText: transpiledComponentText,
+      declaration,
       renderComponentToFilesystem: async (outputPath: string) => {
         await this.renderComponentToFilesystem(transpiledComponentText)(this.fileName)(outputPath);
         if (declaration) {