+++ packages/codegen-ui-react/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.14.2](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.14.1...v2.14.2) (2023-06-06)
+
+### Bug Fixes
+
+- add getOverrideProps to StorageManager ([#1020](https://github.com/aws-amplify/amplify-codegen-ui/issues/1020)) ([8ff282f](https://github.com/aws-amplify/amplify-codegen-ui/commit/8ff282f5a5b2f42dbbb20737d1060a3f5e60cc90))
+
+## [2.14.1](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.14.0...v2.14.1) (2023-05-23)
+
+### Bug Fixes
+
+- fix storage field prop import to storagemanager ([#1016](https://github.com/aws-amplify/amplify-codegen-ui/issues/1016)) ([adf26f4](https://github.com/aws-amplify/amplify-codegen-ui/commit/adf26f43fe6823ba15480a81a2513bb2a98bd7b2))
+
+# [2.14.0](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.13.0...v2.14.0) (2023-05-22)
+
+### Bug Fixes
+
+- **codegen-ui:** fix multiple has one relationships ([#1005](https://github.com/aws-amplify/amplify-codegen-ui/issues/1005)) ([96b7b9d](https://github.com/aws-amplify/amplify-codegen-ui/commit/96b7b9dc3e9e70e1d8527d9abb59334b0bb56ca7))
+- parse operand value when field is number type ([#1013](https://github.com/aws-amplify/amplify-codegen-ui/issues/1013)) ([ad9ba60](https://github.com/aws-amplify/amplify-codegen-ui/commit/ad9ba606ab381b2b4759f19f9310a0901dd10fa8))
+
+### Features
+
+- exporting processFile to be used in studio ([#1001](https://github.com/aws-amplify/amplify-codegen-ui/issues/1001)) ([ff1ef27](https://github.com/aws-amplify/amplify-codegen-ui/commit/ff1ef272dd0b2192217419300353f8b7388aa26c))
+
 ## [2.13.1](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.13.0...v2.13.1) (2023-05-02)
 
 **Note:** Version bump only for package @aws-amplify/codegen-ui-react